data_parsing/requirements.txt
psycopg2
python-crontab